Make objectStoreEquals a shared utility function
